Move matrix exporter
