+++ evaluator/evaluator_threads_manager.py
@@ -1,9 +1,9 @@
 import logging
 
 from config.cst import CONFIG_EVALUATORS_WILDCARD, EvaluatorMatrixTypes, START_PENDING_EVAL_NOTE, \
-    CONFIG_SAVE_EVALUATION, EVALUATION_SAVING_FILE_ENDING, EVALUATION_SAVING_COLUMN_SEPARATOR, \
-    EVALUATION_SAVING_ROW_SEPARATOR
+    CONFIG_SAVE_EVALUATION
 from evaluator.evaluator import Evaluator
+from tools.exporter import MatrixExporter
 
 """
 This class represent the last level of evaluator management by :
@@ -31,13 +31,12 @@ class EvaluatorThreadsManager:
         self.symbol_evaluator = symbol_evaluator
 
         self.should_save_evaluations = CONFIG_SAVE_EVALUATION in self.config and self.config[CONFIG_SAVE_EVALUATION]
-        self.evaluation_save_file = self._get_evaluation_save_file_name()
-        self.is_evaluation_save_file_initialised = False
 
         # notify symbol evaluator
         self.symbol_evaluator.add_evaluator_thread_manager(self.exchange, self.time_frame, self.trading_mode, self)
 
         self.matrix = self.symbol_evaluator.get_matrix(self.exchange)
+        self.matrix_exporter = MatrixExporter(self.matrix, self.symbol)
 
         self.thread_name = f"TA THREAD MANAGER - {self.symbol} - {self.exchange.get_name()} - {self.time_frame}"
         self.logger = logging.getLogger(self.thread_name)
@@ -120,46 +119,7 @@ class EvaluatorThreadsManager:
 
     def _save_evaluations_if_necessary(self):
         if self.should_save_evaluations and self.symbol_evaluator.are_all_timeframes_initialized(self.exchange):
-            try:
-                if not self.is_evaluation_save_file_initialised:
-                    with open(self.evaluation_save_file, 'w+') as eval_file:
-                        eval_file.write(self._get_init_evaluation_save_file())
-                        eval_file.write(self._get_formatted_matrix())
-                else:
-                    with open(self.evaluation_save_file, 'a') as eval_file:
-                        eval_file.write(self._get_formatted_matrix())
-            except PermissionError as e:
-                self.logger.error(f"Impossible to save evaluation on {self.evaluation_save_file}: {e}")
-
-
-    def _get_formatted_matrix(self):
-        formatted_matrix = ""
-        for eval_type, eval_dict in self.matrix.matrix.items():
-            for eval_name, eval_results in eval_dict.items():
-                if isinstance(eval_results, dict):
-                    for eval_result in eval_results.values():
-                        formatted_matrix = f"{formatted_matrix}{EVALUATION_SAVING_COLUMN_SEPARATOR}{eval_result}"
-                else:
-                    formatted_matrix = f"{formatted_matrix}{EVALUATION_SAVING_COLUMN_SEPARATOR}{eval_results}"
-        return formatted_matrix[len(EVALUATION_SAVING_COLUMN_SEPARATOR):] + EVALUATION_SAVING_ROW_SEPARATOR
-
-    def _get_init_evaluation_save_file(self):
-        # create file 1st line
-        line = ""
-        for eval_type, eval_dict in self.matrix.matrix.items():
-            for eval_name, eval_results in eval_dict.items():
-                if isinstance(eval_results, dict):
-                    for eval_key in eval_results.keys():
-                        line = f"{line}{EVALUATION_SAVING_COLUMN_SEPARATOR}" \
-                               f"{eval_type.value}/{eval_name}/{eval_key.value}"
-                else:
-                    line = f"{line}{EVALUATION_SAVING_COLUMN_SEPARATOR}{eval_type.value}/{eval_name}"
-        self.is_evaluation_save_file_initialised = True
-        # remove first separator
-        return line[len(EVALUATION_SAVING_COLUMN_SEPARATOR):] + EVALUATION_SAVING_ROW_SEPARATOR
-
-    def _get_evaluation_save_file_name(self):
-        return f"{self.symbol.replace('/','-')}_{EVALUATION_SAVING_FILE_ENDING}"
+            self.matrix_exporter.save()
 
     def start_threads(self):
         pass
